Directory changes + Docker support
