--- cogs/insultfile.py
@@ -1,27 +1,0 @@
-
-class insult_list():
-    list = [
-        "fuck you",
-        "eat shit",
-        "kill yourself NOW",
-        "placeholder insult because fuck you",
-        "faggot frog",
-        "turkish",
-        "french hippie faggot HON HON HON",
-        "canada is not a real country",
-        "eat my nuts you sad sack of shit",
-        "go fuck yourself",
-        "get off discord and go outside",
-        "Where's your tomboy tradwife, huh?",
-        "You are now manually blinking",
-        "You are now manually breathing",
-        "dry lip having mf DONT LICK THEM FAGGOT",
-        "so down bad you give discord money every month",
-        "if furries want to be animals then my hunting license is valid",
-        "get off reddit",
-        "eat my big dick",
-        "faggot",
-        "you'll never accomplish anything great, kill yourself",
-        "tranny",
-        "you will wake up with furries in your bed tomorrow morning"
-    ]
