+++ config/service.txt
@@ -1,0 +1,7 @@
+botmgr
+pkgmgr
+errorhandler
+basecmd
+cogs.extended
+cogs.fun
+testing.av.ytdl
